refactor generic tooltip
